examples/autogen/pyproject.toml
[project]
name = "py-code-mode-autogen-example"
version = "0.1.0"
description = "AutoGen agent example with py-code-mode"
requires-python = ">=3.12"
dependencies = [
    "py-code-mode[mcp,redis]",
    "autogen-agentchat>=0.4",
    "autogen-ext[anthropic]>=0.4",
    "python-dotenv",
    "tiktoken>=0.12.0",
]

[tool.uv]
dev-dependencies = []

[tool.uv.sources]
py-code-mode = { path = "../..", editable = true }
